requirement.txt
aiohttp==3.8.4
aiosignal==1.3.1

async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cos-python-sdk-v5==1.9.25
crcmod==1.7
databases==0.7.0
fastapi==0.99.1
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
loguru==0.7.0
multidict==6.0.4
pip==23.1.2
psycopg2-binary==2.9.6
pycryptodome==3.18.0
pydantic==1.10.12
pyparsing==3.1.0
requests==2.31.0
setuptools==67.8.0
six==1.16.0
sniffio==1.3.0
sqlalchemy==1.4.48
sqlalchemy_utils==0.41.1
starlette==0.27.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
xmltodict==0.13.0
yarl==1.9.2
pytz==2023.3
pyjwt==1.7.1
asgiref==3.7.2
addressparser==0.2.4

